+++ jt-808-server-support/src/main/java/io/github/hylexus/jt/jt808/support/annotation/msg/resp/ResponseFieldAlias.java
@@ -3,6 +3,7 @@ package io.github.hylexus.jt.jt808.support.annotation.msg.resp;
 import io.github.hylexus.jt.jt808.support.data.MsgDataType;
 import io.github.hylexus.jt.jt808.support.data.serializer.Jt808FieldSerializer;
 import io.github.hylexus.jt.jt808.support.data.serializer.extension.ExtendedJt808FieldSerializerBcdTime;
+import io.github.hylexus.jt.jt808.support.data.serializer.extension.ExtendedJt808FieldSerializerBytes;
 import org.springframework.core.annotation.AliasFor;
 
 import java.lang.annotation.*;
@@ -135,10 +136,25 @@ public @interface ResponseFieldAlias {
         java.lang.String conditionalOn() default "";
 
         @AliasFor(annotation = ResponseField.class, attribute = "customerFieldSerializerClass")
-        Class<? extends Jt808FieldSerializer<?>> customerFieldSerializerClass() default Jt808FieldSerializer.PlaceholderFiledSerializer.class;
+        Class<? extends Jt808FieldSerializer<?>> customerFieldSerializerClass() default ExtendedJt808FieldSerializerBytes.class;
 
         @AliasFor(annotation = ResponseField.class, attribute = "desc")
         java.lang.String desc() default "";
+
+        /**
+         * com.google.common.base.Strings#padEnd(java.lang.String, int, char)
+         */
+        Padding paddingRight() default @Padding(minLength = 0);
+
+        /**
+         * com.google.common.base.Strings#padStart(java.lang.String, int, char)
+         */
+        Padding paddingLeft() default @Padding(minLength = 0);
+
+        /**
+         * 仅仅针对于 {@link java.lang.String} 到 {@link MsgDataType#BYTES} 的场景
+         */
+        java.lang.String charset() default "GBK";
     }
 
     @Target({ElementType.FIELD})