+++ jt-808-server-support/src/main/java/io/github/hylexus/jt/jt808/support/data/MsgDataType.java
@@ -1,6 +1,8 @@
 package io.github.hylexus.jt.jt808.support.data;
 
 import io.github.hylexus.jt.jt808.support.annotation.msg.req.RequestField;
+import io.github.hylexus.jt.jt808.support.data.type.bytebuf.ByteBufContainer;
+import io.github.hylexus.jt.jt808.support.data.type.byteseq.ByteArrayContainer;
 import io.netty.buffer.ByteBuf;
 import lombok.Getter;
 
@@ -14,12 +16,14 @@ import static java.util.Set.of;
  */
 @Getter
 public enum MsgDataType {
-    BYTE(1, "无符号单字节整型(字节，8 位)", of(byte.class, Byte.class, Short.class, short.class, int.class, Integer.class, long.class, Long.class)),
-    BYTES(0, "", of(byte[].class, Byte[].class, ByteBuf.class)),
-    WORD(2, "无符号双字节整型(字，16 位)", of(short.class, Short.class, int.class, Integer.class, long.class, Long.class)),
-    DWORD(4, "无符号四字节整型(双字，32 位)", of(long.class, Long.class, int.class, Integer.class)),
-    BCD(0, "8421 码，n 字节", of(String.class)),
-    STRING(0, "GBK 编码，若无数据，置空", of(String.class)),
+    BYTE(1, "无符号单字节整型(字节，8 位)", of(byte.class, Byte.class, Short.class, short.class, int.class, Integer.class, long.class, Long.class,
+            ByteArrayContainer.class, ByteBufContainer.class)),
+    BYTES(0, "", of(byte[].class, Byte[].class, ByteBuf.class, String.class, ByteArrayContainer.class, ByteBufContainer.class)),
+    WORD(2, "无符号双字节整型(字，16 位)", of(short.class, Short.class, int.class, Integer.class, long.class, Long.class,
+            ByteArrayContainer.class, ByteBufContainer.class)),
+    DWORD(4, "无符号四字节整型(双字，32 位)", of(long.class, Long.class, int.class, Integer.class, ByteArrayContainer.class, ByteBufContainer.class)),
+    BCD(0, "8421 码，n 字节", of(String.class, ByteArrayContainer.class, ByteBufContainer.class)),
+    STRING(0, "GBK 编码，若无数据，置空", of(String.class, ByteArrayContainer.class, ByteBufContainer.class)),
     /**
      * @since 2.0.0
      */