+++ jt-808-server-support/src/main/java/io/github/hylexus/jt/jt808/support/data/serializer/extension/package-info.java
@@ -1,0 +1,8 @@
+
+/**
+ * 提供了一些 <span color="red">特殊</span> 场景下的类型转换器。
+ * <p>
+ * 仅仅用在 <span color="red">特殊</span> 场景下，
+ * <span color="red">不要</span> 注册到 {@link io.github.hylexus.jt.jt808.support.data.serializer.Jt808FieldSerializerRegistry Jt808FieldSerializerRegistry} 中。
+ **/
+package io.github.hylexus.jt.jt808.support.data.serializer.extension;