+++ jt-808-server-support/src/main/java/io/github/hylexus/jt/jt808/support/data/serializer/impl/StringFieldSerializer.java
@@ -1,5 +1,6 @@
 package io.github.hylexus.jt.jt808.support.data.serializer.impl;
 
+import io.github.hylexus.jt.jt808.JtProtocolConstant;
 import io.github.hylexus.jt.jt808.support.data.ConvertibleMetadata;
 import io.github.hylexus.jt.jt808.support.data.MsgDataType;
 import io.github.hylexus.jt.jt808.support.data.ResponseMsgConvertibleMetadata;
@@ -17,7 +18,8 @@ public class StringFieldSerializer implements Jt808FieldSerializer<String> {
 
     private static final Set<ResponseMsgConvertibleMetadata> SUPPORTED_RESPONSE_MSG_CONVERTIBLE_METADATA = Set.of(
             ConvertibleMetadata.forJt808ResponseMsgDataType(String.class, MsgDataType.BCD),
-            ConvertibleMetadata.forJt808ResponseMsgDataType(String.class, MsgDataType.STRING)
+            ConvertibleMetadata.forJt808ResponseMsgDataType(String.class, MsgDataType.STRING),
+            ConvertibleMetadata.forJt808ResponseMsgDataType(String.class, MsgDataType.BYTES)
     );
 
     @Override
@@ -27,7 +29,9 @@ public class StringFieldSerializer implements Jt808FieldSerializer<String> {
 
     @Override
     public void serialize(String object, MsgDataType msgDataType, ByteBuf byteBuf) throws Jt808FieldSerializerException {
-        // TODO NPE
+        if (object == null) {
+            return;
+        }
         switch (msgDataType) {
             case STRING:
                 JtProtocolUtils.writeString(byteBuf, object);
@@ -35,6 +39,10 @@ public class StringFieldSerializer implements Jt808FieldSerializer<String> {
             case BCD:
                 JtProtocolUtils.writeBcd(byteBuf, object);
                 break;
+            case BYTES: {
+                JtProtocolUtils.writeBytes(byteBuf, object.getBytes(JtProtocolConstant.JT_808_STRING_ENCODING));
+                break;
+            }
             default: {
                 throw new Jt808FieldSerializerException("Can not serialize String as " + msgDataType);
             }