+++ jt-808-server-support/src/main/java/io/github/hylexus/jt/jt808/support/data/type/ByteArrayWrapper.java
@@ -6,6 +6,7 @@ import io.netty.buffer.ByteBuf;
 /**
  * @author hylexus
  */
+@Deprecated(since = "2.1.1")
 public class ByteArrayWrapper implements BytesValueWrapper<byte[]> {
     private byte[] value;